# remove offsetWidth call in loop to improve webui performance