offsetWidth is way to expensive to call it for every processbar.

I would be fine with the old 80px processbar. Otherwise we could store the offsetWidth calculation result based on the processbar table header someware as var and use it during td adjustments.

Something about how extrem the performance impact is:

Table with 200 Lines in chrom:
8000ms of 10000ms execution time for "style calculation" at line 864.

Table with 3500 lines
can't profile it but removing offsetWidth reduce the time from 150s to under 10s

But performance during resizing is quite good compared to the performance at initailisation. It does just lag a little bit with offsetWidth during resizing.