@native-api 2017-03-02
Isn't hardcoding a Bad Thing(tm)? One fixed size isn't necessarily appropriate for everyone. I could understand if you rather suggested a single resize at the end instead of one after each line.