@Falcosc 2017-03-02
I did just use the value which was accepted on the 3.3 branch. I mean, fixed size is better then 180s waiting time, or not? :D