@buinsky 2017-03-03
I created the pull request #6447 .

@Falcosc Can you test it?

If resizing of the 'progress' column is too slow it сan be updated after the column resizing (not in real time). Let me know if this necessary.