@Falcosc 2017-03-03
Resizing was quite good for my opinion. I will test your PR tomorrow. Thanks.